메인메뉴에서 0번 선택시 프로그램 종료

+++ src/database/AccountDao.java
@@ -25,12 +25,12 @@ public class AccountDao {
         dbManager.writeAccountFile(accountNumber, lines);
     }
     
-    public int getBalance(String accountNumber) throws IOException {
+    public long getBalance(String accountNumber) throws IOException {
         List<String> lines = dbManager.readAccountFile(accountNumber);
         // Assuming the balance is stored as the first line in the account file
         String balanceStr = lines.get(0);
 
-        return Integer.parseInt(balanceStr);
+        return Long.parseLong(balanceStr);
     }
 
     public String checkAccountNumber(String accountNumber) throws IOException {
@@ -50,7 +50,7 @@ public class AccountDao {
         String balanceStr = lines.get(0);
         DecimalFormat decimalFormat = new DecimalFormat("###,###");
 
-        return decimalFormat.format(Integer.parseInt(balanceStr));
+        return decimalFormat.format(Long.parseLong(balanceStr));
     }
 
     public String getStartDate(String accountNumber, int index) throws IOException {
@@ -74,12 +74,12 @@ public class AccountDao {
             String[] parts = carryBackStr.split("\t");
             DecimalFormat decimalFormat = new DecimalFormat("###,###");
 
-            return decimalFormat.format(Integer.parseInt(parts[0]));
+            return decimalFormat.format(Long.parseLong(parts[0]));
         }
         return accountNumber;
     }
 
-    public void updateBalance(String accountNumber, int newBalance) throws IOException {
+    public void updateBalance(String accountNumber, long newBalance) throws IOException {
         List<String> lines = dbManager.readAccountFile(accountNumber);
         // Update balance in the first line
         lines.set(0, String.valueOf(newBalance));
@@ -109,7 +109,7 @@ public class AccountDao {
     public String showSavings (String accountNumber) throws IOException{
         List<String> lines = dbManager.readAccountFile(accountNumber);
         DecimalFormat decimalFormat = new DecimalFormat("###,###");
-        return decimalFormat.format(Integer.parseInt(lines.get(0)));
+        return decimalFormat.format(Long.parseLong(lines.get(0)));
     }
 
     public boolean hasSavings (String accountNumber, int index) throws IOException{
